Moved list to TodoList component, added AddTodoForm component with form to add todos, and integrated both components in App.jsx
